backend/requirements.txt
fastapi
uvicorn
torch
torch-geometric
numpy
python-multipart
GitPython